feat(metrics): Static metrics registry vite plugin
