+++ src/lib/ctx/assets/vite.ts
@@ -1,0 +1,29 @@
+import { queryAllProjects } from './api.js'
+import { Query } from '../../api/executor.js'
+
+export const fetchAllProjects = () =>
+  queryAllProjects({
+    // @ts-ignore
+    fetcher: globalThis.__VITE_GQL_FETCHER,
+    executor: Query,
+  })().catch((e) => {
+    console.error(e)
+    return []
+  })
+
+export async function StaticAssetsListPlugin() {
+  return {
+    name: 'static-assets-list',
+
+    async transform(src: string, id: string, env: { ssr: boolean }) {
+      if (env.ssr && id.includes('/ctx/assets/index.svelte.')) {
+        const assets = await fetchAllProjects()
+
+        return {
+          code: src.replace('DEFAULT_ASSETS = []', `DEFAULT_ASSETS = ${JSON.stringify(assets)}`),
+          map: null,
+        }
+      }
+    },
+  }
+}
