+++ src/lib/ctx/metrics-registry/index.ts
@@ -1,4 +1,6 @@
-export { useMetricsRegistryCtx, queryGetOrderedMetrics } from './index.svelte.js'
+export { queryGetOrderedMetrics, type TMetricsRegistry, type TRegistryMetric } from './api.js'
+
+export { useMetricsRegistryCtx } from './index.svelte.js'
 
 export { useMetricsRestrictionsCtx, type TMetricRestrictions } from './restrictions/index.svelte.js'
