+++ src/lib/ctx/metrics-registry/vite.ts
@@ -1,0 +1,41 @@
+import { queryGetOrderedMetrics } from './api.js'
+import { Query } from '../../api/executor.js'
+
+export const fetchMetricsRegistry = () =>
+  queryGetOrderedMetrics({
+    // @ts-ignore
+    fetcher: globalThis.__VITE_GQL_FETCHER,
+    executor: Query,
+  })().catch((e) => {
+    console.error(e)
+    return { categories: [], MetricsRegistry: {} }
+  })
+
+export async function StaticMetricsRegistryPlugin() {
+  return {
+    name: 'static-metrics-registry',
+
+    async transform(src: string, id: string, env: { ssr: boolean }) {
+      if (id.includes('/ctx/metrics-registry/index.svelte.')) {
+        const data = await fetchMetricsRegistry()
+
+        let code = src.replace(
+          `DEFAULT_METRIC_CATEGORIES = []`,
+          `DEFAULT_METRIC_CATEGORIES = ${data.categories}`,
+        )
+
+        if (env.ssr) {
+          code = src.replace(
+            'DEFAULT_METRICS_REGISTRY = {}',
+            `DEFAULT_METRICS_REGISTRY = ${JSON.stringify(data)}`,
+          )
+        }
+
+        return {
+          code,
+          map: null,
+        }
+      }
+    },
+  }
+}
